Refactoring of interactive sources -> moved core to pcl -> use c# regex instead of java regex
