+++ HiPMobile/HiPMobile.Droid/Helpers/InteractiveSources/SwitchTabAndScrollToItemInteractiveSourceAction.cs
@@ -15,6 +15,7 @@
 using System;
 using Android.Support.V4.View;
 using Android.Support.V7.Widget;
+using de.upb.hip.mobile.pcl.BusinessLayer.InteractiveSources;
 
 namespace de.upb.hip.mobile.droid.Helpers.InteractiveSources
 {
