+++ HiPMobile/HiPMobile.Droid/Helpers/InteractiveSources/ToastInteractiveSourceAction.cs
@@ -14,6 +14,7 @@
 
 using Android.Content;
 using Android.Widget;
+using de.upb.hip.mobile.pcl.BusinessLayer.InteractiveSources;
 
 namespace de.upb.hip.mobile.droid.Helpers.InteractiveSources {
     /// <summary>
