+++ HiPMobile/HiPMobile/BusinessLayer/InteractiveSources/IInteractiveSourceSubstitute.cs
@@ -1,36 +1,30 @@
 ﻿// Copyright (C) 2016 History in Paderborn App - Universität Paderborn
-//  
+// 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-//  
+// 
 //       http://www.apache.org/licenses/LICENSE-2.0
-//  
+// 
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-namespace de.upb.hip.mobile.droid.Helpers.InteractiveSources {
 
+
+namespace de.upb.hip.mobile.pcl.BusinessLayer.InteractiveSources {
     /// <summary>
-    /// Represents a source by providing properties storing the source text, the start index and the substitute text.
+    /// Uses to determine the substitute that replaces the source markdown 
+    /// of an InteractiveSource
     /// </summary>
-    public class Source {
-        public string SubstituteText { get; }
+    public interface IInteractiveSourceSubstitute {
 
-        public string Text { get; }
+        /// <summary>
+        /// Returns the next substitute.
+        /// </summary>
+        /// <returns>Next substitute string.</returns>
+        string NextSubstitute ();
 
-        public int StartIndex { get; }
-
-        public int NumberInSubtitles { get; }
-
-        public Source(string text, int startIndex, string substituteText, int numberInSubtitles)
-        {
-            Text = text;
-            StartIndex = startIndex;
-            SubstituteText = substituteText;
-            NumberInSubtitles = numberInSubtitles;
-        }
     }
 }
