+++ HiPMobile/HiPMobile/HiPMobile.csproj
@@ -33,6 +33,13 @@
     <ConsolePause>false</ConsolePause>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="BusinessLayer\InteractiveSources\ConsecutiveNumberAndConstantInteractiveSourceSubstitute.cs" />
+    <Compile Include="BusinessLayer\InteractiveSources\ConsecutiveNumberInteractiveSourceSubstitute.cs" />
+    <Compile Include="BusinessLayer\InteractiveSources\ConstantInteractiveSourceSubstitute.cs" />
+    <Compile Include="BusinessLayer\InteractiveSources\IInteractiveSourceAction.cs" />
+    <Compile Include="BusinessLayer\InteractiveSources\IInteractiveSourceSubstitute.cs" />
+    <Compile Include="BusinessLayer\InteractiveSources\InteractiveSourcesParser.cs" />
+    <Compile Include="BusinessLayer\InteractiveSources\Source.cs" />
     <Compile Include="BusinessLayer\Managers\DbManager.cs" />
     <Compile Include="BusinessLayer\Managers\ExhibitManager.cs" />
     <Compile Include="BusinessLayer\Managers\PageManager.cs" />
